move refinedSearch.ts to features/Search
